refactor: move toStoragePath helper function to util/storage/helpers.ts for better organization Fixed issue #5
